fixed bug caused by previous fix + bonus fixes!
Fixing players being invisible + frozen when entering a game in progress caused
the event to be triggered improperly and players were visible and able to move
during event, along with other problems. So that bug is fixed, along with an
endscreen crash caused by new drying code as well as 'level completed' sfx
being played when new players joined a game in progress past level 1.