painting event now happens gradually
Sam now paints the paintings gradually which looks better to the player
and feels more immersive. The system is rather clunky and I will be
making optimizations in the future. Right now the event is handled mostly
server-side but it may be more benefitial for the server to simply send
event triggers and have the events handled client-side. I'll keep testing
and see how it works out best.