candy cane AI smart as fuck now
1. Revamped candy cane AI:
a)If a player is closer to the candy cane than Sam, then Sam "rushes"
to get it with a speed that puts her on fair ground with the
player.
b) If Sam is closer to the candy cane than any player, then she is a
good sport and waits until the players and her are on fair grounds to
get it.
c) Human reaction times are taken into account with above changes.
d) Every 2 times the candy is picked up, the game checks whether sam
or the players have been getting more candy canes.

-If Sam is getting more, the weighted value for human reaction time is
increased so players have more time to react to candy spawning.

-If the players are getting more, the weighted value for human reaction
time is decreased so players have less time to react to candy spawning.

2. Added a data member for speed to the player class which
should be accessed via getSpeed() for all instances of code dealing with
player speed. This will remove speed hardcoding and if we change default
player speed it only needs to be changed in the player class header.

3. Server now calls setVisible() for active/inactive players. Needed this
for candy cane code.