fixed event bug + game start camera jitter
1.Fixed bug where players joining a game in progress after the painting
event still had all the paintings as blank. This can be implemented
more cleanly once we properly handle checking if players are joining
a game already in progress.

2.Fixed camera jitter at start of game caused by calling centercamera()
before scene transition was completed. Not to be confused with camera
jerking from center of window to player location, which still needs
smoothing.

3. Made gradual painting in event a little more gradual.