bucket highlighting implemented
When a player enters the area of a paint bucket where they can
successfully get paint from it, the bucket is now highlighted.
When they leave that area the highlight disappears. This should
resolve the issue of knowing when you're in the 'sweet spot'
and can get paint from a bucket.