camera smoothing + candy cane fix
1. When camera is further than normal from player, camera transitions
over to player more smoothly. I used a common lerping method (that's
a real word somehow) that works really well for all camera use, except
it causes a lot of warping and screen tearing when used currently. If
anyone wants to look into that that and fix it or come up with a better
camera implementation in general, that'd be awesome.

2. Fixed a problem where candy cane pixel coords were not being converted
into their true tile coords, which allowed candy cane to spawn off the map
as well as other problems.