code cleanup + hit sfx fixes
A LOT of code cleanup. And still only the tip of the iceberg. In regard
to code cleanup, major changes are:
1. Player labels are now part of the player class and have been
revamped with superior functionality.
2. Game objects (player,villain,candy cane,etc) are now initialized
in initialize(), which sets variables that would otherwise be set
in client/serverdemo::init(). This cleans up init() a lot.
3. Updated code to not use deprecated cocos2d code anymore, such as
anything prefixed with CC.
4. Cleaned up a lot of unused/unneeded/repetitive variables.
5. Removed a lot of comments that we are unlikely to ever need again.
6. Removed any code related to level 5, since end game screen should
not be considered a level.
7. Cleaned up loadlevel() to be more versatile and replaces a lot of
unnecessary code in init() functions.
8. tilespritevector is now a part of the puzzle class.

AUDIO FIXES:
1. Changed player hit sfx to only happen when a player is hit by
sam/ptero and they erase tiles. Before hit sfx played whenever
player collided with sam/ptero.